Revert "ReBase 4.4.0"
This reverts commit de4561be85f1a8a3c98e9d5b2c0f43912efdb35e.

+++ driver-icarus.h
@@ -136,7 +136,7 @@ struct ICARUS_INFO {
 #ifdef USE_ZEUSMINER
 	// Hardware information, doesn't affect anything directly
 	uint16_t freq;
-	uint16_t chips;
+	uint8_t chips;
 #endif
 };
